refactor: update MpscStream import paths across service modules
